fixed an issue where some edges were skipped
some sample points on the outside of the volume were having neighbour edges incorrectly set to -1, due to an oversight on the lattice structure (on odd layers, some of this outside-edge-rejection cannot be applied due to those edges being in use by tetrahedra later). this resulted in missing triangle indices during the geometry step.